refac v2 WIP
